requirements.txt
flask
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu121
pillow
numpy
opencv-python
imutils
scikit-learn
gunicorn
requests
matplotlib
seaborn
tqdm
albumentations
flask-cors
ultralytics
segmentation-models-pytorch
torch-directml
imutils
